move get_monospace_font_string() to common.py
