+++ sunflower/common.py
@@ -4,6 +4,7 @@ import subprocess
 import locale
 import sys
 
+from gi.repository import Gio, Pango
 
 # user directories
 class UserDirectory:
@@ -37,6 +38,9 @@ class SizeFormat:
 			SI: ['B','kB','MB','GB','TB'],
 			IEC: ['B','KiB','MiB','GiB','TiB']
 		}
+
+
+MONOSPACE_FONT_STRING = None
 
 
 def format_size(size, format_type, include_unit=True):
@@ -216,3 +220,16 @@ def encode_file_name(file_name):
 	"""Encode filename to bytes so it can be passed to GI APIs that expect a file name
 	(and specify `filename` as their argument type in the GIR bindings)."""
 	return str(file_name).encode('utf-8', 'surrogateescape')
+
+def get_monospace_font_string():
+	global MONOSPACE_FONT_STRING
+	if MONOSPACE_FONT_STRING is None:
+		schema = Gio.SettingsSchemaSource.get_default()
+		gnome_interface = schema.lookup('org.gnome.desktop.interface',True)
+		if gnome_interface is None:
+			# not in gnome desktop environment, use 'monospace'
+			MONOSPACE_FONT_STRING = 'monospace'
+		else:
+			settings = Gio.Settings.new('org.gnome.desktop.interface')
+			MONOSPACE_FONT_STRING = settings.get_string('monospace-font-name')
+	return MONOSPACE_FONT_STRING
