+++ sunflower/plugins/file_list/file_list.py
@@ -21,7 +21,6 @@ from sunflower.parameters import Parameters
 from sunflower.plugin_base.item_list import ItemList
 from sunflower.plugin_base.monitor import MonitorSignals, MonitorError
 from sunflower.plugin_base.provider import FileType, Mode as FileMode, Support as ProviderSupport
-from sunflower.tools import system_font
 from sunflower.widgets.thumbnail_view import ThumbnailView
 from sunflower.widgets.emblems_renderer import CellRendererEmblems
 
@@ -325,7 +324,7 @@ class FileList(ItemList):
 		"""Apply font size from settings."""
 		options = self._parent.plugin_options.section(self._name)
 
-		font = system_font.get_monospace_font_string()
+		font = common.get_monospace_font_string()
 
 		for column in columns:
 			column.set_sizing(Gtk.TreeViewColumnSizing.AUTOSIZE)
